% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aide.R
\name{aide}
\alias{aide}
\alias{aide.correlate}
\alias{aide.iqr}
\alias{aide.mean}
\alias{aide.meanSD}
\alias{aide.median}
\alias{aide.medianIQR}
\alias{aide.stddev}
\alias{aide.tertile}
\title{Basic statistics aide}
\usage{
aide.median(x, digits = 1)

aide.iqr(x, digits = 1)

aide.correlate(x, y, digits = 2, method = "spearman",
  na.action = "complete.obs")

aide.medianIQR(x, digits = 1)

aide.mean(x, digits = 1)

aide.stddev(x, digits = 1)

aide.meanSD(x, digits = 1)

aide.tertile(x)
}
\arguments{
\item{x}{vector of values}

\item{digits}{number of digits}

\item{y}{vector of paired values to \code{x}}

\item{method}{the correlation method to use}

\item{na.action}{what to do for missing values, especially for
statistics using paired data (correlation)}
}
\description{
Aide to basic statistics.
}
\details{
Small helper (aide) functions for commonly used basic statistics
functions.
}
\author{
Luke W. Johnston
}

