% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geeglm_confint.R
\name{confint.geeglm}
\alias{confint.geeglm}
\title{Confidence interval for \code{geeglm} objects}
\usage{
\method{confint}{geeglm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{parm}{a specification of which parameters are to be given
    confidence intervals, either a vector of numbers or a vector of
    names.  If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional argument(s) for methods.}
}
\value{
Returns the upper and lower confidence intervals
}
\description{
Generate confidence intervals for GEE analyses
}
\details{
This function was taken from
http://stackoverflow.com/a/21221995/2632184.
}

