% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createCI.R
\name{createCI}
\alias{createCI}
\title{Compute confidence interval for GEE}
\usage{
createCI(data, dist = test.distribution, se = standard.error,
  test.distribution = "Naive.z", standard.error = "Naive.S.E.",
  sig.level = 0.95)
}
\arguments{
\item{data}{The results output from a GEE analysis from the
\code{\link[gee]{gee}} package.}

\item{test.distribution,dist}{The column that contains the test
statistic distribution, e.g. z-score or t-score.}

\item{standard.error,se}{The column that contains the standard
error.}

\item{sig.level}{The significance level for calculating the
confidence interval.}
}
\value{
Columns with the confidence interval and p-value.
}
\description{
Generate confidence intervals (upper and lower) and p-values from
a regression (e.g. GEE or GLM).
}
\examples{
data(state)

## Very simple test example.  Merely to show how the function is used.
outcomes <- c('Income', 'Population')
exposures <- c('Frost', 'Illiteracy')
covariates <- c('Murder', 'LifeExp')
interaction <- 'LifeExp'

## This uses the dplyr package.
ds <- cbind(state.region, state.x77) \%>\%
  as.data.frame() \%>\%
  rename(LifeExp = `Life Exp`,
         ## Need to rename the id variable to SID (see description
         ## above)
         SID = state.region) \%>\%
  arrange(SID)

## Without interaction
loopOutputToListGEE(ds, outcomes, exposures, covariates,
                    corstr = 'exchangeable') \%>\%
  extractBetaFromListGEE() \%>\%
  unlistAndFilterIndep(., exposures) \%>\%
  createCI()

## With interaction
loopOutputToListGEE(ds, outcomes, exposures, covariates,
                    interaction, corstr = 'exchangeable') \%>\%
  extractBetaFromListGEE() \%>\%
  unlistAndFilterIndep(., ':', pattern = TRUE) \%>\%
       ## The ':' represents an interaction
  createCI()
}
\author{
Luke W. Johnston
}

