% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createFormulaList.R
\name{createFormulaList}
\alias{createFormulaList}
\title{Create a list of formulas}
\usage{
createFormulaList(dependent, independent, covariates, interactions = NULL)
}
\arguments{
\item{dependent}{A single or a vector of variables names that the
user wishes to use as dependent variables in an analysis;
\code{dependent} must be as a string/character.}

\item{independent}{Similar to \code{dependent}, except that the
variables would be the independent, or exposure, variables of
interest.}

\item{covariates}{Confounders or covariates that would be included
in \strong{all} formulas and hence \strong{all} models.  Also must
be a string/character.}

\item{interactions}{Optional, include \emph{one} variable from the
\code{covariates} set that will be assigned as an interaction term
with the \code{independent} variable.  Must be a string/character.}
}
\value{
Outputs a list of formulas.
}
\description{
\code{createFormulaList} returns a list of formulas with all
combinations of dependent and independent variables.
}
\details{
This function creates a list of formulas for use in a chain of
processes.  This function's goal is to create all combinations of
a set of dependent, or outcome, variables with a set of
independent, or exposure variables.  Covariates and an interaction
term can also be specified and included into the formula.
}
\examples{
outcomes <- c('Income', 'Education', 'Job')
exposures <- c('Age', 'Sex', 'Height', 'Race', 'IQ')
covariates <- c('ParentEdu', 'Country', 'City')
interactions <- 'City'

createFormulaList(outcomes, exposures, covariates)
createFormulaList(outcomes, exposures, covariates, interactions)
}
\author{
Luke W. Johnston
}

