% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{diagnosticPlots}
\alias{diagnosticPlots}
\title{Regression diagnostic plots and tests}
\usage{
diagnosticPlots(data, y, x, covar)
}
\arguments{
\item{data}{The dataset with the variables of interest}

\item{y}{The dependent or outcome variable (that is, the \code{y}
in the regression equation)}

\item{x}{The independent, exposure, or predictor variable (that
is, the \code{x} in the regression equation)}

\item{covar}{The variables selected as to condition or adjust for
the \code{y} and \code{x} relationship, also known as the
confounding variables}
}
\value{
Outputs multiple plots and textplots with diagnostic
information
}
\description{
Generate regression diagnostic plots and tests for linear
regression models.
}
\details{
This function runs a linear regression on the specified variables
and generates diagnostics based on the regression.  Basic
diagnostics include checking the normality of the residuals,
assessing outliers, influence and Cook's D, and multicollinearity.
Several tests have been commented out, though they can be
uncommented if desired (edit the function to output these if
desired).  Some of the tests I don't fully understand how to
interpret them, but as I learn more I will probably know.  This
function relies on \pkg{MASS} and \pkg{gplots}.
}
\author{
Luke Johnston
}

