% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractBetaFromListGEE.R
\name{extractBetaFromListGEE}
\alias{extractBetaFromListGEE}
\title{Extract beta coefficients from a GEE list object}
\usage{
extractBetaFromListGEE(data)
}
\arguments{
\item{data}{This is variable that contains the list of GEE objects.}
}
\value{
Outputs the beta coefficients.
}
\description{
Used as part of a chain, \code{extractBetaFromListGEE} grabs the
beta estimate from a list of GEE objects.
}
\details{
This is used after \code{\link{loopOutputToListGEE}} as part of a
chain, preferably using dplyr/magrittr's pipe command (%>%).
After generating the list of GEE objects,
\code{extractBetaFromListGEE} loops through each GEE object and
converts the beta estimate and associated statistics into list of
dataframes.
}
\examples{
data(state)

## Very simple test example.  Merely to show how the function is used.
outcomes <- c('Income', 'Population')
exposures <- c('Frost', 'Illiteracy')
covariates <- c('Murder', 'LifeExp')
interaction <- 'LifeExp'

## This uses the dplyr package.
ds <- cbind(state.region, state.x77) \%>\%
  as.data.frame() \%>\%
  rename(LifeExp = `Life Exp`,
         ## Need to rename the id variable to SID (see description
         ## above)
         SID = state.region) \%>\%
  arrange(SID)

## Without interaction
loopOutputToListGEE(ds, outcomes, exposures, covariates,
                    corstr = 'exchangeable') \%>\%
  extractBetaFromListGEE()

## With interaction
loopOutputToListGEE(ds, outcomes, exposures, covariates,
                    interaction, corstr = 'exchangeable') \%>\%
  extractBetaFromListGEE()
}
\author{
Luke W. Johnston
}

