% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractBetaFromListGEE.R
\name{extractBetaGEE}
\alias{extractBetaGEE}
\title{Extract beta coefficients from a GEE object}
\usage{
extractBetaGEE(data)
}
\arguments{
\item{data}{The variable that contains the GEE object.}
}
\description{
Extract the beta object from a GEE object.
}
\author{
Luke W. Johnston
}

