% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{heatmapCorr}
\alias{heatmapCorr}
\title{Correlation heatmap}
\usage{
heatmapCorr(data, x, y, leg.range = c(-1, 1), levels.xlab = NULL,
  levels.ylab = NULL, xlab = "", ylab = "", lo.color = "darkorange2",
  hi.color = "skyblue4", rm.legend = FALSE, matrix.sty = FALSE, ...)
}
\arguments{
\item{data}{Dataset that contains the variables of interest}

\item{x}{Vector of variables that will run along the x-axis}

\item{y}{Same as the \code{x} arg, but for the y-axis}

\item{leg.range}{Range in values for the legend, between -1 and 1}

\item{levels.xlab}{Specifies custom variable names for the x-axis
(it needs to be a list object)}

\item{levels.ylab}{Same as the \code{levels.xlab} arg, but for the y-axis}

\item{xlab}{Sets the label for the x-axis}

\item{ylab}{Same as \code{xlab}, but for the y-axis}

\item{lo.color}{Color of negative correlation coefficients}

\item{hi.color}{Color of positive correlation coefficients}

\item{rm.legend}{In development. Goal is it will remove the legend}

\item{matrix.sty}{In development. Goal is to select whether the
heatmap will be a matrix style (the same variable on the x- and y-
axis) or non-matrix}

\item{...}{Potential options to add. In development}
}
\value{
Outputs a plot object
}
\description{
Generate a matrix or non-matrix style heatmap of correlation
coefficients (i.e. the number of columns and rows can be
different).
}
\details{
This function takes two arguments, the x variables and the y
variables, and generates a heatmap from the variables.  A
correlation matrix is computed from the data, melted (reshape
package), and input into ggplot2 to generate a heatmap.  The output
is the correlations and the plot object.

Dependencies are: \pkg{reshape2} and \pkg{ggplot2}
}
\examples{
xlabs <- list("X Name" = "x1", "X Name Two!" = "x2" ... )
ylabs <- list("Y Name" = "y1", "Y Name Two!" = "y2" ... )
df <- data.frame(replicate(10,sample(0:1,1000,rep=TRUE)))
plot <- heatmap.corr(data = df, x = 1:3, y = 4:10,
                     levels.xlab=xlabs, levels.ylab=ylabs)
print(plot)
}
\author{
Luke Johnston
}

