% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jitterBoxplot}
\alias{jitterBoxplot}
\title{Univariate jittered boxplot}
\usage{
jitterBoxplot(subset.ds, dot.size = 2, dot.colour = "grey50",
  custom.var.names = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{subset.ds}{The dataset that only contains the series of
variables that will be plotted}

\item{dot.size}{Size of the dot for \code{\link[ggplot2]{geom_jitter}}}

\item{dot.colour}{Color of the dots for \code{\link[ggplot2]{geom_jitter}}}

\item{custom.var.names}{List object that contains the custom
(alternative) variable names for the variable (column) names you
passed into the function}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}
}
\value{
Outputs the plot object
}
\description{
Generates a boxplot of variables on one axis with raw values
"jittered" as dots underneath.  The variables need to represent a
similar concept or have the same units for the plot to make sense.
}
\details{
This function is useful for exploring the distribution of a series
of variables that share a common unit, such as kilogram.  The
values for each variable are plotted as jittered dots with a
boxplot of the distribution layered on top of the dots.  The
function takes a subsetted dataset that contains only the series
of variables that share a common unit.  The output object is the
plot.  This function depends on \pkg{ggplot2} and \pkg{reshape2}.
}
\examples{
varnames <- list("Name of X1" = "X1", "Name of X2" = "X2", ...)
df <- subset(ds, VN == 0, select=nefa)
names(df)
jitter.boxplot(df, xlab="Concentration (nmol/mL)",
               ylab="Fatty acid", custom.var.names=varnames)
}
\author{
Luke Johnston
}

