% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loopGEE.R
\name{loopGEE}
\alias{loopGEE}
\title{Loop through multiple generalized estimating equations
(GEE) analyses.}
\usage{
loopGEE(data, dependent, independent, id, covariates = NULL,
  interaction = NULL, corstr = "exchangeable", family = gaussian,
  conf.int = TRUE, conf.level = 0.95, na.rm = TRUE,
  filter.indep = FALSE, filter.interact = FALSE, adjust.p.value = FALSE)
}
\arguments{
\item{data}{Dataset to run GEE on}

\item{dependent}{The dependent (aka outcome or response)
variables.  Must be quoted and can have several.}

\item{independent}{Like \code{dependent}, except the explanatory
(aka predictor or exposure) variables.}

\item{id}{The variable to cluster on for GEE, for instance the
'ID' variable for a person in a longitudinal cohort.}

\item{covariates}{The covariate variables.  Can be multiple
covariates.}

\item{interaction}{A single interaction variable.}

\item{corstr}{a character string specifying the correlation structure. The
          following are permitted: '"independence"', '"exchangeable"',
          '"ar1"', '"unstructured"' and '"userdefined"' }

\item{family}{See corresponding documentation to \code{glm}}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{na.rm}{Remove missing values from the dataset before running
GEE.  \code{geeglm} can't handle any missingness, so sometimes
it's necessary to remove missingness.}

\item{filter.indep}{Logical; Keep only the rows that have the
\code{independent} variables.}

\item{filter.interact}{Logical; Keep only the rows that have the
\code{interaction} variable.}

\item{adjust.p.value}{Logical; Adjust for multiple comparisons
using the False Discovery Rate.}
}
\value{
A dataframe of all the GEE analyses, with estimates,
confidence intervals, and p-values.
}
\description{
Loop through each combination of dependent and independent
variables, running a generalized estimating equations model on
each combination, and generate a dataframe of the results.
}
\examples{
data(state)
ds <- data.frame(state.region, state.x77)
names(ds)
outcome <- c('Income', 'Murder')
exposure <- c('Population', 'Life.Exp', 'Illiteracy')
covar <- c('Area', 'Frost')
cid <- 'state.region'

loopGEE(ds, outcome, exposure, cid)
loopGEE(ds, outcome, exposure, cid, covariates = covar)
loopGEE(ds, outcome, exposure, cid, covariates = covar, interaction = 'HS.Grad')
loopGEE(ds, outcome, exposure, cid, covariates = covar, interaction = 'Area')
loopGEE(ds, outcome, exposure, cid, corstr = 'ar1', conf.level = 0.99)
loopGEE(ds, outcome, exposure, cid, corstr = 'ar1', conf.int = FALSE)
loopGEE(ds, outcome, exposure, cid, corstr = 'ar1', adjust.p.value = TRUE)
loopGEE(ds, outcome, exposure, cid, covariates = covar, adjust.p.value = TRUE)
loopGEE(ds, outcome, exposure, cid, covariates = covar, adjust.p.value = TRUE,
        filter.indep = TRUE)
loopGEE(ds, outcome, exposure, cid, interaction = 'Area', covariates = covar,
        adjust.p.value = TRUE, filter.interact = TRUE)
}
\author{
Luke W. Johnston
}
\seealso{
\code{\link[broom]{tidy}}
}

