% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiplot}
\alias{multiplot}
\title{Multiple plots on page}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Where the ggplot objects are placed to be laid out on
the graph grid}

\item{plotlist}{Can be used in place of the ... argument by
specifying the ggplot objects as a list object}

\item{file}{Not sure what this is used for}

\item{cols}{Number of columns for the layout.  For example cols=2
provides two columns and with four ggplot objects, the resulting
output would be a 2 by 2 graphic}

\item{layout}{A matrix that indicates the plot grid layout.  For
example, if layout = matrix(c(1, 2, 3, 3), nrow = 2, byrow = TRUE)
the result would have plot 1 in the upper left, plot 2 in the upper
right, and plot 3 would be go across the bottom}
}
\description{
Lay out multiple ggplots on one frame or page.
}
\details{
This function, which was from
\url{http://www.cookbook-r.com/Graphs}, is used to lay out several
ggplot objects onto one frame or pdf page.  For instance, you can
have 3 plots on a page, one going vertically across the top, the
other two in each corner on the bottom.  This function makes up
for the difficulty \pkg{ggplot2} has with outputting multiple
plots on one grid.  This function depends on \pkg{grid}.
}
\author{
Cookbook R
}

