% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBoxWithJitter.R
\name{plotBoxWithJitter}
\alias{plotBoxWithJitter}
\title{Jittered-dot boxplot}
\usage{
plotBoxWithJitter(data, x, y, groups = NULL)
}
\arguments{
\item{data}{The dataset with the variables of interest}

\item{x}{Variable for the x-axis}

\item{y}{Variable for the y-axis}

\item{groups}{Optional, split boxplots and jittered dots by group.}
}
\description{
Generates a boxplot of a factor variable on one axis with raw
values "jittered" as dots underneath the box.
}
\details{
This function is useful for exploring the distribution of a series
of variables that share a common unit, such as kilogram.  The
values for each variable are plotted as jittered dots with a
boxplot of the distribution layered on top of the dots.  Can add
axis labels, themes, etc, through \code{\link[ggplot2]{ggplots}}
interface (eg. \code{ylab}).
}
\examples{
data(state)
cbind(state.region, state.x77) \%>\%
  as.data.frame() \%>\%
  mutate(Region = as.factor(state.region),
         MedianIncome = cut(Income, breaks = c(-Inf, median(Income), Inf),
                            labels = c('Lower', 'Higher'))) \%>\%
  plotBoxWithJitter(., 'Region', 'Population', groups = 'MedianIncome')
}
\author{
Luke W. Johnston
}

