% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotHeatmapCorr.R
\name{plotHeatmapCorr}
\alias{plotHeatmapCorr}
\title{Correlation heatmap}
\usage{
plotHeatmapCorr(data, x, y = NULL, x.name.sub, y.name.sub,
  heat.colours = c("darkorange2", "skyblue4"), show.corr.values = TRUE,
  ylab = y.axis.label, xlab = x.axis.label, print.corr.values = FALSE,
  y.axis.label = NULL, x.axis.label = NULL)
}
\arguments{
\item{data}{The dataset to plot.}

\item{x}{The x axis variables.}

\item{y}{The y axis variables.}

\item{x.name.sub}{The substitutions to the x axis names, incase
the original variables need to be clarified (eg. 'Wgt' to
'Weight').  Use (probably) best command to use is the \code{gsub}
command.}

\item{y.name.sub}{Same as the \code{x.name.sub}, but for the y
axis variables.}

\item{heat.colours}{The spectrum of colours for the heat map, as a
vector between the lowest (negative) correlation and the highest
(positive) correlation.}

\item{show.corr.values}{Logical; add the correlation values to the
heatmap.}

\item{print.corr.values}{Logical; If true, prints the correlation values.}

\item{y.axis.label,ylab}{}

\item{x.axis.label,xlab}{}

\item{corr.text.size}{Size of the text that show the correlation
values.}
}
\description{
Generate a matrix or non-matrix style heatmap of correlation
coefficients.
}
\details{
This function takes two arguments, the x variables and the y
variables, and generates a heatmap from the variables.  A
correlation matrix is computed from the data, melted (reshape
package), and input into ggplot2 to generate a heatmap.  The
output is the correlations and the plot object.
}
\examples{
data(state)
plotHeatmapCorr(state.x77, c('Income', 'Population'), c('Frost', 'Murder'))

xvars <- c('Income', 'Population', 'Area', 'Frost')
## Replace Area with 'Land Area (km^2)', etc, on the axes using the gsub command
plotHeatmapCorr(state.x77, xvars, print.corr.values = TRUE,
                x.name.sub = xvars \%>\%
                  gsub('Area', 'Land Area (km^2)', .) \%>\%
                  gsub('Income', 'Income ($)', .)
                )
}
\author{
Luke W. Johnston
}

