% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSmoothingANOVA.R
\name{plotSmoothingANOVA}
\alias{plotSmoothingANOVA}
\title{Smoothing line plot, with ANOVA}
\usage{
plotSmoothingANOVA(data, x, y, id, y.axis.limits = c(0.2, 0.8))
}
\arguments{
\item{data}{Dataset with the variables of interest.}

\item{x}{The factor variable on the x-axis.  Generally is the time
variable in a longitudinal setting.  Must be a string/character.}

\item{y}{The continuous variable on the y-axis.  Must be a
string/character.}

\item{id}{The grouping variable, generally the ID for the
participant in a longitudinal setting.}

\item{y.axis.limits}{Limits of the y-axis.  Must be two numbers.}
}
\description{
Generate a bivariate plot of a factor variable (x-axis) and a
continuous variable (y-axis) that overlays a smoothing line
(loess), which also prints an ANOVA p-value.
}
\details{
Generally I use this for plotting longitudinal data, where the
x-axis is the timepoints and the ANOVA is testing the significance
across time (which may be arguably inappropriate... FIXME).
}
\examples{
data(state)
cbind(state.region, state.x77) \%>\%
  as.data.frame() \%>\%
  mutate(f.Pop = Population \%>\%
      quantile(., c(0, .333, .666, 1), na.rm = TRUE) \%>\%
      cut(Population, ., include.lowest = TRUE)) \%>\%
  plotSmoothingANOVA(., 'f.Pop', 'Illiteracy', id = 'state.region')
}
\author{
Luke W. Johnston
}

