% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSpaghetti.R
\name{plotSpaghetti}
\alias{plotSpaghetti}
\title{Spaghetti plot}
\usage{
plotSpaghetti(data, y, x, groups = "SID")
}
\arguments{
\item{data}{The dataset to plot.}

\item{y}{The variable to go on the y-axis.}

\item{x}{The variable to go on the x-axis.}

\item{groups}{The unique ID variable to differentiate subjects in
a longitudinal dataset.}
}
\description{
Plot subjects in a longitudinal dataset, making a 'spaghetti'
plot.
}
\examples{
\dontrun{
## A pretend case (not a real example)
plotSpaghetti(dataset, 'Height', 'Year', 'SubjectID')
}
}
\author{
Luke W. Johnston
}

