% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{plotVisreg}
\alias{plotVisreg}
\title{Visualizing adjusted linear regression models}
\usage{
plotVisreg(data, y, x, covar, ylabel = x, xlabel = y, ...)
}
\arguments{
\item{data}{Dataset with the variables of interest}

\item{y}{The dependent or outcome variable in the regression
equation}

\item{x}{The independent or exposure variable in the regression
equation}

\item{covar}{The confounding variables, that is the variables being adjusted for}

\item{ylabel}{The y-axis label}

\item{xlabel}{The x-axis label}

\item{...}{Other options. In development}
}
\value{
Outputs a plot of the regression model
}
\description{
Generates plots of a linear regression model which includes
confounding variables.
}
\details{
This function runs a linear regression on the specified variables
and plots the partial residuals.  This allows for visualizing the
relationship between the outcome and the exposure, after adjusting
for confounders.  A linear slope is plotted through the partial
residuals, with a confidence interval band around it.  The output
is a plot.  This function depends on \pkg{visreg}.
}
\author{
Luke Johnston
}

