% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/smoothPlot.R
\name{smoothPlot}
\alias{smoothPlot}
\title{Smooth plot}
\usage{
smoothPlot(x, y, data, ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{data}{}

\item{...}{}
}
\value{
Outputs a plot
}
\description{
In development ..
}
\details{
.. content for details ..
}
\author{
Luke Johnston
}

