% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/smoothTimePlot.R
\name{smoothTimePlot}
\alias{smoothTimePlot}
\title{Smooth time plots}
\usage{
smoothTimePlot(dsn, x, y, id, smooth.type = "loess", ylab = y, xlab = x)
}
\arguments{
\item{dsn}{The dataset name.}

\item{x}{Specify the variable for the x-axis.  Must be
either in quotes or as a number.}

\item{y}{As with \code{x}, but for the y-axis.}

\item{id}{Specify the ID variable for longitudinal or repeated
measures type data.}

\item{smooth.type}{Set the smoothing method, which includes
"loess", "lm", "rlm".}

\item{ylab}{Optionally set the y-axis label.}

\item{xlab}{Optionally set the x-axis label.}
}
\value{
Outputs a ggplot object
}
\description{
Generates a plot of two variables, particularly for longitudinal
data.  The x-axis is typically the time variable.
}
\details{
This function creates a plot with smooth lines, typically using
the loess method, along with the means of a discrete time variable
(eg. time 0, 1, 2, 3 etc).

Dependencies are: \pkg{ggplot2}
}
\author{
Luke Johnston
}

