% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{summarySE}
\alias{summarySE}
\title{Summarize means and standard errors of the mean}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{A dataset (dataframe) that contains the values to be summarized}

\item{measurevar}{The name of a column that contains the variable to be summarized}

\item{groupvars}{A vector containing names of columns that contain grouping variables}

\item{na.rm}{A binary (boolean) response that indicates whether to
ignore missing (NA) data}

\item{conf.interval}{Percent range of the confidence interval}
}
\value{
Outputs a dataframe that contains the summarized statistics (means, etc.)
}
\description{
Calculates the sample size, mean, standard deviation, standard
error of the mean, and the confidence interval of specified
variables.
}
\details{
I took this function on 2014-01-21 from the website
\url{http://www.cookbook-r.com/Graphs}.  It basically summarizes
the provided data by giving count, mean, standard deviation,
standard error of the mean, and confidence interval (default 95%).
The dependencies are \pkg{plyr}
}
\author{
Cookbook R
}

