% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{theme_white}
\alias{theme_white}
\title{Custom white ggplot theme}
\usage{
theme_white()
}
\description{
Creates a white, simple theme for \pkg{ggplot2} objects
}
\details{
The default \pkg{ggplot2} theme is decent for most purposes, but
is visually unappealing.  This function aims to correct that by
setting the theme to something more similar to the default in the
base R plot package.  The function dependes on \pkg{ggplot2}.
}
\examples{
## This creates a white theme
theme_white()
}
\author{
Luke Johnston
}

