% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unlistAndFilterIndep.R
\name{unlistAndFilterIndep}
\alias{unlistAndFilterIndep}
\title{Unlist a list of dataframes and filter a pattern or variable.}
\usage{
unlistAndFilterIndep(data, x, pattern = FALSE)
}
\arguments{
\item{data}{The list object with the dataframes}

\item{x}{The variables of interest (eg. exposures) that are within
the 'indep' column.}

\item{pattern}{Logical: TRUE if \code{x} is a pattern rather than
an explicit list of variables and FALSE if \code{x} is an explicit
list of variables (eg. exposures).}
}
\value{
Outputs a single dataframe with only the rows with the
variables from \code{x}.
}
\description{
Unlist a list of dataframes, convert into a single dataframe, and
filter out a string or pattern from the 'indep' column.
}
\details{
This function is generally used within a chain of other commands
that creates a list of dataframes, generally from a regression.
The list gets unlisted and converted into a single dataset.
Afterward, the dataset gets filtered by the variables of interest
(eg. the exposures) that are contained within the 'indep' column.
}
\examples{
## Very simple test example.  Merely to show how the function is used.
outcomes <- c('Income', 'Population')
exposures <- c('Frost', 'Illiteracy')
covariates <- c('Murder', 'LifeExp')
interaction <- 'LifeExp'

## This uses the dplyr package.
ds <- cbind(state.region, state.x77) \%>\%
  as.data.frame() \%>\%
  rename(LifeExp = `Life Exp`,
         ## Need to rename the id variable to SID (see description
         ## above)
         SID = state.region) \%>\%
  arrange(SID)

## Without interaction
loopOutputToListGEE(ds, outcomes, exposures, covariates,
                    corstr = 'exchangeable') \%>\%
  extractBetaFromListGEE() \%>\%
  unlistAndFilterIndep(., exposures)

## With interaction
loopOutputToListGEE(ds, outcomes, exposures, covariates,
                    interaction, corstr = 'exchangeable') \%>\%
  extractBetaFromListGEE() \%>\%
  unlistAndFilterIndep(., ':', pattern = TRUE)
       ## The ':' represents an interaction
}
\author{
Luke W. Johnston
}

