% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loopGEE.R
\name{loopGEE}
\alias{loopGEE}
\title{Loop through multiple GEE analyses.}
\usage{
loopGEE(data, dependent, independent, id, covariates = NULL,
  interaction = NULL, corstr = "exchangeable", family = gaussian,
  conf.int = TRUE, conf.level = 0.95)
}
\arguments{
\item{data}{Dataset to run GEE on}

\item{dependent}{The dependent (aka outcome or response)
variables.  Must be quoted and can have several.}

\item{independent}{Like \code{dependent}, except the explanatory
(aka predictor or exposure) variables.}

\item{id}{The variable to cluster on for GEE, for instance the
'ID' variable for a person in a longitudinal cohort.}

\item{covariates}{The covariate variables.  Can be multiple
covariates.}

\item{interaction}{A single interaction variable.}

\item{corstr}{a character string specifying the correlation structure. The
          following are permitted: '"independence"', '"exchangeable"',
          '"ar1"', '"unstructured"' and '"userdefined"' }

\item{family}{See corresponding documentation to \code{glm}}
}
\value{
A dataframe of all the GEE analyses, with estimates,
confidence intervals, and p-values.
}
\description{
Loop through each combination of dependent and independent
variables and generate a dataframe of the results.
}
\examples{
data(state)
ds <- data.frame(state.region, state.x77)
loopGEE(ds, c('Income', 'Frost'), c('Population', 'Murder'), 'state.region')
loopGEE(ds, 'Income', 'Population', 'state.region',
covariates = c('Frost', 'Area'))
loopGEE(ds, 'Income', 'Population', 'state.region',
covariates = 'Frost', interaction = 'Frost')
loopGEE(ds, 'Income', 'Population', 'state.region', corstr = 'ar1',
conf.int = FALSE)
}
\author{
Luke W. Johnston
}
\seealso{
\code{\link[broom]{tidy}}
}

