% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{smoothBivarPlot}
\alias{smoothBivarPlot}
\title{Smooth bivariate plot}
\usage{
smoothBivarPlot(x, y, data, ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{data}{}

\item{...}{}
}
\value{
Outputs a plot
}
\description{
In development ..
}
\details{
.. content for details ..
}
\author{
Luke Johnston
}

